Fix warning in rust wrapper about `env::home_dir`.
