+++ wrappers/rust/tests/payment.rs
@@ -3,9 +3,9 @@ extern crate rust_libindy_wrapper as indy;
 #[macro_use] extern crate serde_derive;
 extern crate rmp_serde;
 extern crate byteorder;
+
 #[macro_use]
 mod utils;
-
 use utils::wallet::Wallet;
 
 mod low_tests {
