+++ wrappers/rust/tests/utils/constants.rs
@@ -4,10 +4,8 @@ pub const DID: &str = "8wZcEriaNLNKtteJvx7f8i";
 pub const DID_TRUSTEE: &str = "V4SGRU86Z58d6TV7PBUe6f";
 pub const INVALID_HANDLE: i32 = 583741;
 pub const METADATA: &str = "some_metadata";
-pub const MY1_SEED: &str = "00000000000000000000000000000My1";
 pub const PROTOCOL_VERSION: i32 = 2;
 pub const SEED_1: &str = "00000000000000000000000000000My1";
-pub const TRUSTEE_SEED: &str = "000000000000000000000000Trustee1";
 pub const VERKEY_1: &str = "GjZWsBLgZCR18aL468JAT7w9CZRiBnpxUPPgyQxh4voa";
 pub const VERKEY_ABV_1: &str = "~HYwqs2vrTc8Tn4uBV7NBTe";
 pub const VERKEY_TRUSTEE: &str = "GJ1SzoWzavQYfNL9XkaJdrQejfztN4XqdsiV4ct3LXKL";
