+++ wrappers/rust/tests/utils/environment.rs
@@ -1,5 +1,6 @@
+extern crate dirs;
+
 use std::env;
-use std::fs;
 use std::path::PathBuf;
 
 #[cfg(test)]
@@ -7,59 +8,51 @@ pub fn test_pool_ip() -> String {
     env::var("TEST_POOL_IP").unwrap_or("127.0.0.1".to_string())
 }
 
-pub fn tmp_path() -> PathBuf {
-    let mut path = env::temp_dir();
-    path.push("indy_client");
-    path
-}
-
-pub fn tmp_file_path(file_name: &str) -> PathBuf {
-    let mut path = tmp_path();
-    path.push(file_name);
-    path
-}
-
-pub fn pool_home_path() -> PathBuf {
+pub fn pool_path(pool_name: &str) -> PathBuf {
     let mut path = indy_home_path();
     path.push("pool");
-    path
-}
-
-pub fn pool_path(pool_name: &str) -> PathBuf {
-    let mut path = pool_home_path();
     path.push(pool_name);
     path
 }
 
+#[cfg(not(any(target_os = "android", target_os = "ios")))]
 pub fn indy_home_path() -> PathBuf {
-    // TODO: FIXME: Provide better handling for the unknown home path case!!!
-    let mut path = env::home_dir().unwrap_or(PathBuf::from("/home/indy"));
-    let mut indy_client_dir = ".indy_client";
-    if cfg!(target_os = "ios"){
-        indy_client_dir = "Documents/.indy_client";
-    }
-    path.push(indy_client_dir);
-
-    if cfg!(target_os = "android"){
-        android_create_indy_client_dir();
-        path = android_indy_client_dir_path();
-    }
+    let mut path = dirs::home_dir().unwrap();
+    path.push(".indy_client");
     path
 }
-pub fn android_indy_client_dir_path() -> PathBuf{
-    let external_storage= env::var("EXTERNAL_STORAGE");
-    let android_dir :String;
-    match external_storage {
-        Ok (val) => android_dir = val + "/.indy_client",
-        Err(err) => {
-            panic!("Failed to find external storage path {:?}", err)
-        }
-    }
 
-    PathBuf::from(android_dir)
+#[cfg(target_os = "ios")]
+pub fn indy_home_path() -> PathBuf {
+    let mut path = dirs::home_dir().unwrap();
+    path.push("Documents/.indy_client");
+    path
 }
 
-pub fn android_create_indy_client_dir(){
-    //Creates directory only if it is not present.
-    fs::create_dir_all(android_indy_client_dir_path().as_path()).unwrap();
+#[cfg(target_os = "android")]
+pub fn indy_home_path() -> PathBuf {
+    android_create_indy_client_dir();
+    android_indy_client_dir_path()
+}
+
+#[cfg(target_os = "android")]
+pub mod android {
+    use std::fs;
+
+    pub fn indy_client_dir_path() -> PathBuf {
+        let external_storage = env::var("EXTERNAL_STORAGE");
+        let android_dir = match external_storage {
+            Ok (val) => val + "/.indy_client",
+            Err(err) => {
+                panic!("Failed to find external storage path {:?}", err)
+            }
+        };
+
+        PathBuf::from(android_dir)
+    }
+
+    pub fn create_indy_client_dir() {
+        //Creates directory only if it is not present.
+        fs::create_dir_all(indy_client_dir_path().as_path()).unwrap();
+    }
 }
